Add files via upload
Hi buddies. This is the project which is belong to ROBOTIC PROCESS AUTOMATAION (RPA) domain which  automates the resume generation which is basically a bot which automates the process .This is done by using the Uipath Studiox and can be done using Uipath Studio also. So here the source code for the Resume generation. Have a look at it.